:recycle: #114 GitHubリポジトリをfetchする部分をカスタムフックに切り出し
